refactor(fuzz): drive pub(crate) logic via a feature-gated in-crate harness
Stop exposing internal types as pub just for fuzzing. The harness logic for
velocity_enforce / effective_balance / ec_rollup_batch now lives in feature-
gated #[cfg(feature = "fuzz")] __fuzz submodules inside their owning modules
(where the types/methods are already visible), and is re-exported through a
top-level #[doc(hidden)] cala_ledger::fuzz.

Reverts all the visibility promotions this PR had added:
- lib.rs: pub mod ec_rollup -> mod ec_rollup
- balance: pub mod effective -> mod effective; EffectiveBalanceData /
  SnapshotOrEntry back to pub(super); EcRollupTxn back to pub(crate)
- ec_rollup: EcRollupBatch / PendingTx + their methods back to private
- velocity/account_control: pub use value::* -> pub(super) use value::*;
  dropped the pub re-exports of the enforcement types and EvalContext

Default (production) builds now expose nothing extra — the __fuzz modules and
cala_ledger::fuzz don't even compile without the feature. The empty-input
guard fix in re_calculate_snapshots is retained.

The out-of-tree fuzz crate enables cala-ledger/fuzz and its three targets
become one-liners that feed bytes to cala_ledger::fuzz::*.

Verified: cargo build -p cala-ledger (default) and --features fuzz both build
warning-free; make fuzz runs all 9 targets clean; effective_balance seed run
unchanged (cov 284).
