+++ cala-ledger/src/velocity/mod.rs
@@ -21,16 +21,6 @@ pub use control::*;
 use error::*;
 pub use limit::*;
 
-// The enforcement snapshot types and the context builder are exposed so that
-// fuzzing/tooling can drive the pure enforcement logic (`needs_enforcement`,
-// `window_for_enforcement`, `enforce`) without a database. They are already
-// `Serialize`/`Deserialize` (they are persisted), so this is a minimal surface
-// expansion rather than a new data model.
-pub use account_control::{
-    AccountBalanceLimit, AccountLimit, AccountVelocityControl, AccountVelocityLimit,
-};
-pub use context::EvalContext;
-
 #[derive(Clone)]
 pub struct Velocities {
     limits: VelocityLimitRepo,
@@ -276,3 +266,57 @@ impl Velocities {
         Ok(self.controls.find_all(control_ids).await?)
     }
 }
+
+#[cfg(feature = "fuzz")]
+mod __fuzz {
+    //! Harness for the out-of-tree `velocity_enforce` fuzz target. Lives in
+    //! this module so it can reach the `pub(super)` enforcement types.
+    use super::account_control::AccountVelocityControl;
+    use super::context::EvalContext;
+    use cala_types::{
+        balance::BalanceSnapshot,
+        entry::EntryValues,
+        transaction::TransactionValues,
+        velocity::VelocityContextAccountValues,
+    };
+    use es_entity::clock::Clock;
+
+    pub fn fuzz_enforce(data: &[u8]) {
+        let parts: Vec<&[u8]> = data.split(|&b| b == 0xFF).collect();
+        if parts.len() < 5 {
+            return;
+        }
+        let Ok(control) = serde_json::from_slice::<AccountVelocityControl>(parts[0]) else {
+            return;
+        };
+        let Ok(entry) = serde_json::from_slice::<EntryValues>(parts[1]) else {
+            return;
+        };
+        let Ok(snapshot) = serde_json::from_slice::<BalanceSnapshot>(parts[2]) else {
+            return;
+        };
+        let Ok(tx) = serde_json::from_slice::<TransactionValues>(parts[3]) else {
+            return;
+        };
+        let Ok(account) =
+            serde_json::from_slice::<VelocityContextAccountValues>(parts[4])
+        else {
+            return;
+        };
+
+        // The account must be registered before we ask for its entry context
+        // (context_for_entry expects it, panicking otherwise).
+        let account_id = account.id;
+        let mut eval = EvalContext::new(Clock::handle().clone(), &tx, std::iter::once(&account));
+        let ctx = eval.context_for_entry(account_id, &entry);
+
+        let _ = control.needs_enforcement(&ctx);
+        for limit in &control.velocity_limits {
+            let _ = limit.window_for_enforcement(&ctx, &entry);
+            let _ = limit.enforce(&ctx, snapshot.created_at, &snapshot);
+        }
+    }
+}
+
+#[cfg(feature = "fuzz")]
+pub use __fuzz::fuzz_enforce;
